version 1.0.0

+++ Restaurants.Application/Dishes/Commands/DeleteDish/DeleteDishCommandHandler.cs
@@ -1,8 +1,10 @@
 ﻿using AutoMapper;
 using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDish;
@@ -10,7 +12,8 @@ namespace Restaurants.Application.Dishes.Commands.DeleteDish;
 internal class DeleteDishCommandHandler(ILogger<DeleteDishCommandHandler> logger,
     IMapper mapper,
     IRestaurantsRepository restaurantsRepository,
-    IDishesRepository dishesRepository) : IRequestHandler<DeleteDishCommand>
+    IDishesRepository dishesRepository,
+    IRestaurantAuthorizationService restaurantAuthorizationService) : IRequestHandler<DeleteDishCommand>
 {
     public async Task Handle(DeleteDishCommand request, CancellationToken cancellationToken)
     {
@@ -20,7 +23,12 @@ internal class DeleteDishCommandHandler(ILogger<DeleteDishCommandHandler> logger
         if (restaurant == null)
         {
             throw new NotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
+        }
 
+
+        if (!restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Update))
+        {
+            throw new ForbidException();
         }
 
         await dishesRepository.Delete(restaurant.Dishes);
